(BSR)[API] chore: move/reorganize external attributes code
- avoid most of the code of "external" module in __init__
- move "external" out of "users" because pro attributes, zendesk, also
  sync info about offerers and venues
- split automations into users and venues
