+++ api/src/pcapi/routes/native/v1/authentication.py
@@ -5,10 +5,10 @@ from flask_jwt_extended import get_jwt_identity
 from flask_jwt_extended import jwt_required
 
 from pcapi.connectors import api_recaptcha
+from pcapi.core.external.attributes import api as external_attributes_api
 from pcapi.core.subscription.dms import api as dms_subscription_api
 from pcapi.core.users import api as users_api
 from pcapi.core.users import exceptions as users_exceptions
-from pcapi.core.users import external as users_external
 from pcapi.core.users import repository as users_repo
 from pcapi.core.users.models import TokenType
 from pcapi.core.users.models import User
@@ -150,7 +150,7 @@ def validate_email(body: ValidateEmailRequest) -> ValidateEmailResponse:
 
     user.isEmailValidated = True
     repository.save(user)
-    users_external.update_external_user(user)
+    external_attributes_api.update_external_user(user)
 
     try:
         dms_subscription_api.try_dms_orphan_adoption(user)
